+++ src/main/java/io/burt/jmespath/contrib/function/MultiplyFunction.java
@@ -1,6 +1,6 @@
 package io.burt.jmespath.contrib.function;
 
-public class MultipleFunction extends MathBiFunction {
+public class MultiplyFunction extends MathBiFunction {
     @Override
     protected double performMathOperation(double x, double y) {
         return x * y;
